requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
sqlalchemy==2.0.23
aiosqlite==0.19.0
paho-mqtt==1.6.1
apscheduler==3.10.4
aiohttp==3.9.1
pyyaml==6.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
